+++ src/main/java/com/aetherwars/model/card/character/Attackable.java
@@ -6,5 +6,5 @@ import com.aetherwars.model.player.Player;
 public interface Attackable {
     void attackCharacter(SummonedCharacter c) throws CardException;
     void attackPlayer(Player p);
-    void takeDamage(int damage);
+    void takeDamage(double damage);
 }
